Changes to support bootstrap version 3
Modified the way the version of bootstrap is detected. Ensured that all the 'version 2' icon classes are still added, incase fontawesome or some other glyph library is being used (and hasn't added support for version 3, yet.  Added code to trigger a change on 'hide' if the value has changed. This is necessary in the rare event that someone opens the datepicker, and closes it without selecting a date/time. The input get populated with the current date, but no event was being triggered.